refactor: delete_checkpoint
